:wrench:

+++ example/src/Examples/Breathe/demo3.tsx
@@ -1,6 +1,6 @@
 export const demo3 = async(device: GPUDevice, context: GPUCanvasContext) => {
    // Create an array of data
-   const data = new Float32Array([1.0, 2.0, 3.0, 4.0]);
+   const data = new Uint32Array([1.0, 2.0, 3.0, 4.0]);
 
    // Create a GPU buffer and store data
    const gpuBuffer = device.createBuffer({
@@ -16,7 +16,7 @@ export const demo3 = async(device: GPUDevice, context: GPUCanvasContext) => {
   await gpuBuffer.mapAsync(1, 0, data.byteLength); // GPUMapMode.READ
   console.log("mapAsync done");
   const arrayBuffer = gpuBuffer.getMappedRange(0,  data.byteLength);
-  const readData = new Float32Array(arrayBuffer);
+  const readData = new Uint32Array(arrayBuffer);
 
   //  // Output the read data
   console.log('Data read from the buffer:', readData);
